hiic-hr-app/backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
pandas>=1.5.3
numpy>=1.24.2
supabase>=1.0.3
httpx>=0.24.0
python-multipart>=0.0.6
jinja2>=3.1.2
markdown>=3.4.3
langchain>=0.1.0 